fix(vault-encryption): satisfy clippy for Sub-D doc + i128 conversion
Sub-D 工程3 fix-up #1.

CI 初回 lint で 4 errors:
- mod.rs L7/L9/L11: rustdoc `[`error`]:` 形式 link reference defined in list
  item (rustdoc::link_reference_in_list は clippy::all level deny)。リスト
  項目内の link reference 定義は rendered ドキュメントで非表示になるため
  禁止。バッククォート区切り `error` / `header` / `record` / `recovery` /
  `confirmation` / `service` / `storage` の単純テキスト表記に切替。
- header.rs L227: `i128::from(self.created_at.unix_timestamp_nanos())` が
  clippy::useless_conversion で reject。`time::OffsetDateTime::
  unix_timestamp_nanos()` は既に i128 を返すため i128::from() は冗長、削除
  してコメントで根拠記録。

ローカル `cargo fmt --all -- --check`: PASS。
sub-d-static-checks.sh: 3/4 PASS (S03 SKIP は Sub-F 設計通り、変動なし)。

Refs #42

+++ crates/shikomi-infra/src/persistence/vault_migration/header.rs
@@ -224,7 +224,8 @@ impl VaultEncryptedHeader {
 
         // created_at の i64 マイクロ秒変換 (Aad と同じ規約)。
         // 範囲外時はゼロ埋めで防衛的フォールバック (構造的に到達しない、time crate 限界外)。
-        let micros = i128::from(self.created_at.unix_timestamp_nanos()) / 1_000;
+        // unix_timestamp_nanos() は既に i128 を返すため i128::from() は冗長 (clippy::useless_conversion)。
+        let micros = self.created_at.unix_timestamp_nanos() / 1_000;
         let micros_i64 = i64::try_from(micros).unwrap_or(0);
         out.extend_from_slice(&micros_i64.to_be_bytes());
